python/requirements.txt
Django==1.11.7
gunicorn==19.7
mysqlclient==1.3.12
setuptools==28.8.0
pytz==2017.3
selenium==3.8.0
pyvirtualdisplay==0.2.1
pysolr==3.7.0
django-haystack==2.6.0